refactor: move shared notebook launching code into a hook
